backend/task_2/requirement.txt
requests
pydub
praat-parselmouth
numpy
python-dotenv